# Add simple benchmark test