--- a/consentconstants/errors.go
+++ b/consentconstants/errors.go
@@ -1,0 +1,8 @@
+package consentconstants
+
+import "errors"
+
+var (
+	// ErrEmptyDecodedConsent error raised when the consent string is empty
+	ErrEmptyDecodedConsent = errors.New("decoded consent cannot be empty")
+)