--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,5 @@
+module github.com/prebid/go-gdpr
+
+go 1.13
+
+require github.com/buger/jsonparser v0.0.0-20180318095312-2cac668e8456