--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,7 @@ module github.com/prebid/go-gdpr
 
 go 1.13
 
-require github.com/buger/jsonparser v0.0.0-20180318095312-2cac668e8456
+require (
+	github.com/buger/jsonparser v0.0.0-20180318095312-2cac668e8456
+	github.com/stretchr/testify v1.7.0
+)