--- a/vendorconsent/tcf1/metadata.go
+++ b/vendorconsent/tcf1/metadata.go
@@ -9,6 +9,10 @@ import (
 	"github.com/prebid/go-gdpr/consentconstants"
 )
 
+var (
+  errInvalidVendorListVersion = errors.New("the consent string encoded a VendorListVersion of 0, but this value must be greater than or equal to 1")
+)
+
 // Parse the metadata from the consent string.
 // This returns an error if the input is too short to answer questions about that data.
 func parseMetadata(data []byte) (consentMetadata, error) {
@@ -23,7 +27,7 @@ func parseMetadata(data []byte) (consentMetadata, error) {
 		return nil, fmt.Errorf("the consent string encoded a Version of %d, but this value must be greater than or equal to 1", metadata.Version())
 	}
 	if metadata.VendorListVersion() == 0 {
-		return nil, errors.New("the consent string encoded a VendorListVersion of 0, but this value must be greater than or equal to 1")
+		return nil, errInvalidVendorListVersion
 
 	}
 	return consentMetadata(data), nil