--- a/vendorconsent/tcf2/metadata.go
+++ b/vendorconsent/tcf2/metadata.go
@@ -56,6 +56,7 @@ type pubRestrictResolver interface {
 	CheckPubRestriction(purposeID uint8, restrictType uint8, vendor uint16) bool
 }
 
+// Version returns the version stored in the first 5 bits
 func (c ConsentMetadata) Version() uint8 {
 	// Stored in bits 0-5
 	return uint8(c.data[0] >> 2)
@@ -66,6 +67,7 @@ const (
 	decisPerOne  = 10
 )
 
+// Created returns the created date stored in bits 6 to 41
 func (c ConsentMetadata) Created() time.Time {
 	// Stored in bits 6-41.. which is [000000xx xxxxxxxx xxxxxxxx xxxxxxxx xxxxxxxx xx000000] starting at the 1st byte
 	deciseconds := int64(binary.BigEndian.Uint64([]byte{
@@ -81,6 +83,7 @@ func (c ConsentMetadata) Created() time.Time {
 	return time.Unix(deciseconds/decisPerOne, (deciseconds%decisPerOne)*nanosPerDeci)
 }
 
+// LastUpdated returns the last updated date stored in bits 42 to 77
 func (c ConsentMetadata) LastUpdated() time.Time {
 	// Stored in bits 42-77... which is [00xxxxxx xxxxxxxx xxxxxxxx xxxxxxxx xxxxxx00 ] starting at the 6th byte
 	deciseconds := int64(binary.BigEndian.Uint64([]byte{
@@ -96,6 +99,7 @@ func (c ConsentMetadata) LastUpdated() time.Time {
 	return time.Unix(deciseconds/decisPerOne, (deciseconds%decisPerOne)*nanosPerDeci)
 }
 
+// CmpID returns the Consent Management Platform identifier stored in bits 78 to 89
 func (c ConsentMetadata) CmpID() uint16 {
 	// Stored in bits 78-89... which is [000000xx xxxxxxxx xx000000] starting at the 10th byte
 	leftByte := ((c.data[9] & 0x03) << 2) | c.data[10]>>6
@@ -103,6 +107,7 @@ func (c ConsentMetadata) CmpID() uint16 {
 	return binary.BigEndian.Uint16([]byte{leftByte, rightByte})
 }
 
+// CmpVersion returns the Consent Management Platform version stored in bits 90 to 101
 func (c ConsentMetadata) CmpVersion() uint16 {
 	// Stored in bits 90-101.. which is [00xxxxxx xxxxxx00] starting at the 12th byte
 	leftByte := (c.data[11] >> 2) & 0x0f
@@ -110,11 +115,13 @@ func (c ConsentMetadata) CmpVersion() uint16 {
 	return binary.BigEndian.Uint16([]byte{leftByte, rightByte})
 }
 
+// ConsentScreen returns the consent screen info stored in bits 102 to 107
 func (c ConsentMetadata) ConsentScreen() uint8 {
 	// Stored in bits 102-107.. which is [000000xx xxxx0000] starting at the 13th byte
 	return uint8(((c.data[12] & 0x03) << 4) | c.data[13]>>4)
 }
 
+// ConsentLanguage returns the two letter code for consent language stored in bits 108 to 119
 func (c ConsentMetadata) ConsentLanguage() string {
 	// Stored in bits 108-119... which is [0000xxxx xxxxxxxx] starting at the 14th byte.
 	// Each letter is stored as 6 bits, with A=0 and Z=25
@@ -127,6 +134,7 @@ func (c ConsentMetadata) VendorLegitInterestMaxID() uint16 {
 	return c.vendorLegitimateInterests.MaxVendorID()
 }
 
+// VendorListVersion returns the vendor list version stored in bits 120 to 131
 func (c ConsentMetadata) VendorListVersion() uint16 {
 	// The vendor list version is stored in bits 120 - 131
 	rightByte := ((c.data[16] & 0xf0) >> 4) | ((c.data[15] & 0x0f) << 4)
@@ -134,6 +142,7 @@ func (c ConsentMetadata) VendorListVersion() uint16 {
 	return binary.BigEndian.Uint16([]byte{leftByte, rightByte})
 }
 
+// MaxVendorID returns the maximum value for vendor identifier in bits 213 to 228 
 func (c ConsentMetadata) MaxVendorID() uint16 {
 	// The max vendor ID is stored in bits 213 - 228 [00000xxx xxxxxxxx xxxxx000]
 	leftByte := ((c.data[26] & 0x07) << 5) | ((c.data[27] & 0xf8) >> 3)
@@ -141,15 +150,17 @@ func (c ConsentMetadata) MaxVendorID() uint16 {
 	return binary.BigEndian.Uint16([]byte{leftByte, rightByte})
 }
 
+// PurposeAllowed returns if the given purpose (1 to 24 max) is enabled, info stored in bits 152 to 175
 func (c ConsentMetadata) PurposeAllowed(id consentconstants.Purpose) bool {
 	// Purposes are stored in bits 152 - 175. The interface contract only defines behavior for ints in the range [1, 24]...
 	// so in the valid range, this won't even overflow a uint8.
 	if id > 24 {
-		id = 24
+		id = 24 // tiago: why this behavior is different than PurposeLITransparency? should we check for 0 and negative too? 
 	}
 	return isSet(c.data, uint(id)+151)
 }
 
+// PurposeLITransparency returns if the given purpose transparency (1 to 24 max) is enabled, info stored in bits 176 to 199
 func (c ConsentMetadata) PurposeLITransparency(id consentconstants.Purpose) bool {
 	// Purposes are stored in bits 176 - 199. The interface contract only defines behavior for ints in the range [1, 24]...
 	// so in the valid range, this won't even overflow a uint8.
@@ -158,11 +169,12 @@ func (c ConsentMetadata) PurposeLITransparency(id consentconstants.Purpose) bool
 	}
 	return isSet(c.data, uint(id)+175)
 }
-
+// PurposeOneTreatment returns if Purpose 1 is enable, info stored in bit 200
 func (c ConsentMetadata) PurposeOneTreatment() bool {
 	return isSet(c.data, 200)
 }
 
+// SpecialFeatureOptIn returns if the given special feature is enable, stored in bits 139 to 151
 func (c ConsentMetadata) SpecialFeatureOptIn(id uint16) bool {
 	if id > 12 {
 		return false
@@ -170,14 +182,17 @@ func (c ConsentMetadata) SpecialFeatureOptIn(id uint16) bool {
 	return isSet(c.data, 140+uint(id)-1)
 }
 
+// VendorConsent returns true if there is consent for the given vendor id
 func (c ConsentMetadata) VendorConsent(id uint16) bool {
 	return c.vendorConsents.VendorConsent(id)
 }
 
+// VendorLegitInterest returns true if there is legitimate interest established for the given vendor id
 func (c ConsentMetadata) VendorLegitInterest(id uint16) bool {
 	return c.vendorLegitimateInterests.VendorConsent(id)
 }
 
+// CheckPubRestriction returns the publisher restriction for a given purpose id, restriction type and vendor id
 func (c ConsentMetadata) CheckPubRestriction(purposeID uint8, restrictType uint8, vendor uint16) bool {
 	return c.publisherRestrictions.CheckPubRestriction(purposeID, restrictType, vendor)
 }