--- a/vendorlist/vendorlist.go
+++ b/vendorlist/vendorlist.go
@@ -6,10 +6,12 @@ import (
 
 // Copying from API for backwards compatibility
 
+// VendorList is a backward compatible interface of api.VendorList
 type VendorList interface {
 	api.VendorList
 }
 
+// Vendor is a backward compatible interface of api.Vendor
 type Vendor interface {
 	api.Vendor
 }