Hello

Since this project has no benchmark tests, I decide to contribute with a small one.

I often do this kind of benchmark:

1. I use a different package to run my tests and benchmarks (with an `_test` suffix ) to test only the public interfaces/types/etc. If needed I write some test with suffix`_internal_test.go` to be possible identify quickly.

2. For a first benchmark I just choose some consent strings (bad encoded, v1 and v2) and check what happens if we parse each consent for a big period of time. Also I test all consents in a big "all" test, to mimic the real scenario (multiple kinds of different consents)

old is the v0.8.2
new is this fork

```
$ benchcmp old.txt new.txt 
benchmark                                                          old ns/op     new ns/op     delta
BenchmarkParse/all_testcases-4                                     2311          2143          -7.27%
BenchmarkParse/case_empty_string-4                                 291           3.63          -98.75%
BenchmarkParse/case_explicit_nok_string-4                          239           127           -46.86%
BenchmarkParse/case_bad_encoded_string-4                           217           116           -46.54%
BenchmarkParse/case_short_consent_v1_ok-4                          332           224           -32.53%
BenchmarkParse/case_long_consent_v1_ok-4                           629           428           -31.96%
BenchmarkParse/case_consent_v1_nok-4                               857           713           -16.80%
BenchmarkParse/case_consent_v1_maxvendorid=0-4                     552           425           -23.01%
BenchmarkParse/case_short_consent_v2_ok-4                          895           720           -19.55%
BenchmarkParse/case_long_consent_v2_ok-4                           913           718           -21.36%
BenchmarkParse/case_bad_consent_v2-4                               17450         17350         -0.57%
BenchmarkVerify/testing_just_parsing_10000_consents/string-4       733           519           -29.20%

benchmark                                                          old allocs     new allocs     delta
BenchmarkParse/all_testcases-4                                     6              4              -33.33%
BenchmarkParse/case_empty_string-4                                 3              0              -100.00%
BenchmarkParse/case_explicit_nok_string-4                          3              2              -33.33%
BenchmarkParse/case_bad_encoded_string-4                           3              2              -33.33%
BenchmarkParse/case_short_consent_v1_ok-4                          3              2              -33.33%
BenchmarkParse/case_long_consent_v1_ok-4                           4              2              -50.00%
BenchmarkParse/case_consent_v1_nok-4                               9              7              -22.22%
BenchmarkParse/case_consent_v1_maxvendorid=0-4                     4              3              -25.00%
BenchmarkParse/case_short_consent_v2_ok-4                          10             8              -20.00%
BenchmarkParse/case_long_consent_v2_ok-4                           10             8              -20.00%
BenchmarkParse/case_bad_consent_v2-4                               11             9              -18.18%
BenchmarkVerify/testing_just_parsing_10000_consents/string-4       4              2              -50.00%

benchmark                                                          old bytes     new bytes     delta
BenchmarkParse/all_testcases-4                                     6806          6710          -1.41%
BenchmarkParse/case_empty_string-4                                 112           0             -100.00%
BenchmarkParse/case_explicit_nok_string-4                          88            40            -54.55%
BenchmarkParse/case_bad_encoded_string-4                           88            40            -54.55%
BenchmarkParse/case_short_consent_v1_ok-4                          160           96            -40.00%
BenchmarkParse/case_long_consent_v1_ok-4                           416           224           -46.15%
BenchmarkParse/case_consent_v1_nok-4                               288           192           -33.33%
BenchmarkParse/case_consent_v1_maxvendorid=0-4                     208           144           -30.77%
BenchmarkParse/case_short_consent_v2_ok-4                          442           330           -25.34%
BenchmarkParse/case_long_consent_v2_ok-4                           442           330           -25.34%
BenchmarkParse/case_bad_consent_v2-4                               65820         65708         -0.17%
BenchmarkVerify/testing_just_parsing_10000_consents/string-4       419           231           -44.87%
```

as you can see, we can reduce memory, object allocation and time spent on each benchmark, in trivial cases it is good, in some case is not that great.

I made two kinds of optimization:
- move immutable error object to global variables, this way we create just one error and reuse the same instance (not possible in all cases, since we add some info like the bit related to the error...)
- change how we decode the consent string, if tcfv2 instead split by `.` I just peek the substring until the `.` instead split in until 4 new strings without reason

of course, you may want to parse the the other sections of the tcfv2 consent string in the future, but you can follow the same pattern of index and offset

Other changes:

I run golint and golangci-lint on this code with default options. I note some public methods had no documentation and we have some unused code. 

I add the missing doc and remove the unused code (when it is for tests I just ignore)

I add some TODO to be possible verify another optimization: change some methods to pointer receiver.

also we can reuse the bitutils package in tcfv1 rangesection (but this will change the error messages, I don't want to change it without discussion)

Regards

Tiago

edit: I add one more benchmark test ( to read a file with gdpr unique strings, one per line ) to test with real strings and with this I perform other small modifications.

I see the split per `.` only make sense for V2 strings, and V2 strings starts with C (becase base64) si the ParseString will delegate to tcf1 or tcf2 based on the first char of the consent string. This way we split with dot only if we really need, since tcf1 strings does not need it.

I use a small trick to reduce object allocation in base64 decoding: we can reuse the original byte slice instead create a new one, doing `Decode(buffer,buffer)` and since the destination slice will have less bytes than the source, this is safe.

Since sometimes I need parse a batch of consent strings, each optimization may help me.