@peczenyj 2021-03-01
Sure :)

Em seg., 1/03/2021 às 19:16, Scott Kay <notifications@github.com> escreveu:

> @peczenyj <https://github.com/peczenyj> Very sorry that we haven't looked
> at this in a long time. Would you be interested if we picked it back up for
> review?
>
> —
> You are receiving this because you were mentioned.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/prebid/go-gdpr/pull/23#issuecomment-788161729>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAAOQLXNPR2JSIFNG6E7KNTTBPKZRANCNFSM4NBTBJYQ>
> .
>
-- 
Tiago B. Peczenyj

http://about.me/peczenyj