@peczenyj 2021-03-04
Since TCF v1 will die soon... I don’t think we should spend more time on
this. However with some free time I can send a second PR

Em qui., 4/03/2021 às 17:04, hhhjort <notifications@github.com> escreveu:

> *@hhhjort* commented on this pull request.
> ------------------------------
>
> In vendorconsent/tcf1/rangesection.go
> <https://github.com/prebid/go-gdpr/pull/23#discussion_r587597563>:
>
> > @@ -88,7 +88,7 @@ func parseException(data consentMetadata, initialBit uint) (rangeException, uint
>  }
>
>  // parseUInt16  parses a 16-bit integer from the data array, starting at the given index
> -func parseUInt16(data []byte, bitStartIndex uint) (uint16, error) {
> +func parseUInt16(data []byte, bitStartIndex uint) (uint16, error) { // tiago: why not use the bitutils.ParseUint16 ?
>
> I believe this is here becasue bitutils was written when TCF2 was added,
> and not too much effort was put in to back porting it to the TCF1 code.
> Feel free to back port it in, on the other hand tcf1 should not be seeing
> any use anymore.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/prebid/go-gdpr/pull/23#pullrequestreview-604269761>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAAOQLUGIY7SLZ6QHVYNSMDTB6VO7ANCNFSM4NBTBJYQ>
> .
>
-- 
Tiago B. Peczenyj

http://about.me/peczenyj