@peczenyj 2021-03-04
Sounds good, I can send another PR with this soon™️

Em qui., 4/03/2021 às 17:02, hhhjort <notifications@github.com> escreveu:

> *@hhhjort* commented on this pull request.
> ------------------------------
>
> In vendorconsent/consent.go
> <https://github.com/prebid/go-gdpr/pull/23#discussion_r587595880>:
>
> >
>  	"github.com/prebid/go-gdpr/api"
>  	tcf1 "github.com/prebid/go-gdpr/vendorconsent/tcf1"
>  	tcf2 "github.com/prebid/go-gdpr/vendorconsent/tcf2"
>  )
>
> +var (
> +	errEmptyDecodedConsent = fmt.Errorf("decoded consent cannot be empty")
>
> We use this in several places. Perhaps we want to consider exporting this
> from consentconstants, which would have the side effect of allowing the
> user of the library to do a simple compare to check that an error is this
> error rather than doing string matching.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/prebid/go-gdpr/pull/23#pullrequestreview-604267412>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAAOQLV5HBYN3UA7R6M56A3TB6VIFANCNFSM4NBTBJYQ>
> .
>
-- 
Tiago B. Peczenyj

http://about.me/peczenyj