@peczenyj 2021-03-05
Hi guys I try to address all issues here, if I forget something please let me know.

I also merge with master