@peczenyj 2021-03-05
Do you want me change it or just remove the comments and it can be solved
in another task?

Em sex., 5/03/2021 às 18:07, hhhjort <notifications@github.com> escreveu:

> *@hhhjort* commented on this pull request.
> ------------------------------
>
> In vendorconsent/tcf2/metadata.go
> <https://github.com/prebid/go-gdpr/pull/23#discussion_r588480627>:
>
> >  func (c ConsentMetadata) PurposeAllowed(id consentconstants.Purpose) bool {
>  	// Purposes are stored in bits 152 - 175. The interface contract only defines behavior for ints in the range [1, 24]...
>  	// so in the valid range, this won't even overflow a uint8.
>  	if id > 24 {
> -		id = 24
> +		id = 24 // tiago: why this behavior is different than PurposeLITransparency? should we check for 0 and negative too?
>
> I am not sure how it became inconsistent. We should probably make it
> consistent. The >24 check is more important than != 0, as the bit
> associated with 0 definitely exists, but we may overrun the consent string
> for an ID high enough. Still not a bad idea to include it.
>
> Also a good point it may be better to just return a false for out of
> bounds, it is a bit more predictable.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/prebid/go-gdpr/pull/23#discussion_r588480627>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAAOQLTGQVYSCGTCU5XGATDTCEFVTANCNFSM4NBTBJYQ>
> .
>
-- 
Tiago B. Peczenyj

http://about.me/peczenyj