@SyntaxNode 2021-03-08
> Do you want me change it or just remove the comments and it can be solved in another task?

Either is fine with me. Whichever you feel like.