@peczenyj 2021-03-08
Ok

Em seg., 8/03/2021 às 07:47, Scott Kay <notifications@github.com> escreveu:

> Do you want me change it or just remove the comments and it can be solved
> in another task?
>
> Either is fine with me. Whichever you feel like.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/prebid/go-gdpr/pull/23#issuecomment-792514374>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAAOQLVQFSLWMELX2YZRJ53TCRXIVANCNFSM4NBTBJYQ>
> .
>
-- 
Tiago B. Peczenyj

http://about.me/peczenyj