@peczenyj 2021-03-08
Hello guys, I am a little bit lost on all comments, I try to solve all discussions but perhaps I miss something, please let me know if there is something else to do :)

BTW there is a new version of github.com/buger/jsonparser ( v1.1.1 ) not sure if you want me to update the vendor or perhaps this can be done in the next release

Regards