@SyntaxNode 2021-03-11
> BTW there is a new version of github.com/buger/jsonparser ( v1.1.1 ) not sure if you want me to update the vendor or perhaps this can be done in the next release

Please save that for a future PR. Let's lock in the scope of this PR now to focus on getting in merge. I promise we'll keep up well moving forward with future PRs if you'd like to keep working on this project. :)