@peczenyj 2021-03-12
> Very nice. I just have 2 more comments:
> 
> * In the benchmark test file you're assigning variable names to the result of ParseString call, but then throwing it away. Consider not assigning to the variables in the first place.

Hello

I use this pattern on all my benchmarks

if you look at https://dave.cheney.net/2013/06/30/how-to-write-benchmarks-in-go

section "A note on compiler optimisations"

I am just using a different way to prevent the compilar to affect the benchmark itself. If you insist I can change it but I can't see any advantage


> * I think there are some comments which still have an off-by-one error on the bit count, such as VendorListVersion and MaxVendorID. Could you please review the bit range comments one more time?

Ok I will review it. Thanks