@SyntaxNode 2021-03-12
> I use this pattern on all my benchmarks
> 
> if you look at https://dave.cheney.net/2013/06/30/how-to-write-benchmarks-in-go

Ah, ok. No need to change it. Could you please add a comment explaining the technique? The code looks strange for those unfamiliar with this trapping, such as myself.